Install & update zstandard explicitly
The build produced by GitHub actions is not working for me on Ubuntu due
to an error like this:
```
ImportError: zstd C API versions mismatch
```

As python-zstandard bundles its own copy of zstandard, this mismatch
should not occur. I suspect it's caused by a system version of
python-zstandard (with debundled zstandard) getting in the way, so we force an upgrade here.